Add files via upload
Game mechanics: 
You are a zombie chasing humans, each die is a human victim. There are 13 dies in the bag. You will choose and roll 3 of them at a time with 3 possible outcomes... 
BRAIN - set aside, you have eaten a brain. 
FOOTPRINT - return to the bag, the victim got away 
SHOTGUN - set aside, collect 3 and you are dead, if you die your turn ends and nothing is added to your score. 

The die are colour-coded depending on their probability of rolling a brain, green = high probability, yellow = low probability, red = very low probability. As the probability of rolling a brain decreases the probability of rolling a shotgun increases. 

Whoever collects the most brains wins. This game is Multi-Player.